requirements.txt
python-telegram-bot[job-queue]==21.6
requests>=2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
lxml==5.3.0
aiosqlite>=0.19.0
